Load configuration from args too
args is now a first-class source of configuration along file and environment
variables. This allow to override simple values from argv.
